Contribute componer only with documentBrowser component
