+++ UDV-Core/examples/DemoBrowser.html
@@ -13,6 +13,6 @@
     <script src="../node_modules/moment/moment.js"></script>
     <script src="../node_modules/three/build/three.js"></script>
     <script src="../src/Externals/dat.gui/dat.gui.min.js"></script>
-    <script src="Vilo3D2018.js"></script>
+    <script src="DemoBrowser.js"></script>
 </body>
 </html>
