+++ UDV-Core/src/Modules/Contribute/Contribute.css
@@ -376,7 +376,7 @@ margin: 10px;
     z-index: 10;
     margin : auto;
     bottom : 20%;
-    left: 22%;;
+    right: 5%;
     border: 1px solid black;
     background-color: white;
     color: black;
